+++ internal/resource/common/authenticationpolicytreenode/authentication_policy_tree_node_client_struct.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	client "github.com/pingidentity/pingfederate-go-client/v1200/configurationapi"
+	client "github.com/pingidentity/pingfederate-go-client/v1210/configurationapi"
 	"github.com/pingidentity/terraform-provider-pingfederate/internal/resource/common/policyaction"
 	internaltypes "github.com/pingidentity/terraform-provider-pingfederate/internal/types"
 )