+++ internal/resource/config/authenticationpolicies/settings/authentication_policies_settings_common.go
@@ -2,7 +2,7 @@ package authenticationpoliciessettings
 
 import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	client "github.com/pingidentity/pingfederate-go-client/v1200/configurationapi"
+	client "github.com/pingidentity/pingfederate-go-client/v1210/configurationapi"
 	"github.com/pingidentity/terraform-provider-pingfederate/internal/resource/common/id"
 )